requirements.txt
dapr-agents>=0.8.1
starlette==0.47.2
python-dotenv>=1.0.0
